pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avm"
version = "1.0.0"
description = "Audio to Video Maker - Convert narrated lessons into polished YouTube videos"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AVM Team", email = "team@avm.dev"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=0.10,<0.12",
    "moviepy>=2.0,<3.0",
    "playwright>=1.47,<1.60",
    "markdown-it-py>=3.0,<4.0",
    "jinja2>=3.1,<4.0",
    "PyYAML>=6.0,<7.0",
    "pydantic>=2.7,<3.0",
    "srt>=3.5,<4.0",
    "Pillow>=10.2,<12.0",
    "numpy>=1.26,<2.0",
    "structlog>=24.1,<25.0",
    "typer>=0.12,<0.13",
    "pydub>=0.25,<1.0",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.18,<1.20",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.scripts]
avm = "avm:main"

[project.urls]
Homepage = "https://github.com/avm-team/avm"
Repository = "https://github.com/avm-team/avm"
Issues = "https://github.com/avm-team/avm/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["avm*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
